+++ src/components/datagrid/data_grid.tsx
@@ -14,7 +14,10 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { VariableSizeGrid as Grid } from 'react-window';
+import {
+  VariableSizeGrid as Grid,
+  GridOnItemsRenderedProps,
+} from 'react-window';
 import { useGeneratedHtmlId, keys } from '../../services';
 import { EuiFocusTrap } from '../focus_trap';
 import { EuiI18n, useEuiI18n } from '../i18n';
@@ -172,6 +175,7 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = (props) => {
   // Imperative handler passed back by react-window - we're setting this at
   // the top datagrid level to make passing it to other children & utils easier
   const gridRef = useRef<Grid | null>(null);
+  const gridItemsRendered = useRef<GridOnItemsRenderedProps | null>(null);
 
   /**
    * Display
@@ -257,9 +261,10 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = (props) => {
   const { headerIsInteractive, handleHeaderMutation } = useHeaderIsInteractive(
     contentRef.current
   );
-  const { focusProps: wrappingDivFocusProps, ...focusContext } = useFocus(
-    headerIsInteractive
-  );
+  const { focusProps: wrappingDivFocusProps, ...focusContext } = useFocus({
+    headerIsInteractive,
+    gridItemsRendered,
+  });
 
   /**
    * Toolbar & full-screen
@@ -401,6 +406,8 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = (props) => {
                 gridElement: contentRef.current,
                 visibleColCount,
                 visibleRowCount,
+                visibleRowStartIndex:
+                  gridItemsRendered.current?.visibleRowStartIndex || 0,
                 rowCount,
                 pagination,
                 hasFooter: !!renderFooterCellValue,
@@ -440,6 +447,7 @@ export const EuiDataGrid: FunctionComponent<EuiDataGridProps> = (props) => {
                 gridStyles={gridStyles}
                 gridWidth={gridWidth}
                 gridRef={gridRef}
+                gridItemsRendered={gridItemsRendered}
                 wrapperRef={contentRef}
               />
             </div>